fix shopping cart issue

+++ LOG4420_TP5/client/src/app/shopping-cart/shopping-cart.component.ts
@@ -45,12 +45,14 @@ export class ShoppingCartComponent implements OnInit {
   }
 
   deleteProduct(productId: number): void {
-    const self = this;
-    this.shoppingCartService.deleteProduct(productId)
-      .then(() => {
-        self.products = self.products.filter(p => p.productId != productId ? p : null);
-        self.updateSum();
-      });
+    if(confirm("Voulez-vous supprimerle produit du panier ?")) {
+      const self = this;
+      this.shoppingCartService.deleteProduct(productId)
+        .then(() => {
+          self.products = self.products.filter(p => p.productId != productId ? p : null);
+          self.updateSum();
+        });
+    }
   }
 
   flushCart(): void {
